feat(ARC-3814): move theme group section to separate component
https://meemoo.atlassian.net/browse/ARC-3814
